+++ Lab4/Model/ProductWarehouse.cs
@@ -6,10 +6,10 @@ public class ProductWarehouse
 {
     [Key]
     public int IdProductWarehouse { get; set; }
-    
+
     [Required]
     public int IdWarehouse { get; set; }
-    
+
     [Required]
     public int IdProduct { get; set; }
 
@@ -18,11 +18,11 @@ public class ProductWarehouse
 
     [Required]
     public int Amount { get; set; }
-    
+
     [Required]
     [MaxLength(200)]
-    public float Price { get; set; }
-    
+    public decimal Price { get; set; }
+
     [Required]
     [DataType(DataType.DateTime)]
     public DateTime CreatedAt { get; set; }
